+++ source/mdfour.h
@@ -31,6 +31,6 @@ typedef struct mdfour {
 } mdfour_t;
 
 void mdfour_begin( struct mdfour *md );
-void mdfour_update( struct mdfour *md, uint8_t *in, int n );
+void mdfour_update( struct mdfour *md, uint8_t *in, size_t n );
 void mdfour_result( struct mdfour *md, uint8_t *out );
